improve cli help message and cleaned up argument parser
help() function's docstring was indented back. textwrap.dedent()
was used to display it properly. The description of the argument
parser was set to help()'s docstring and RawDescriptionHelpFormatter
was used to display it without wrapping. The argument parser was
moved out class to a global scope. Imports were sorted
alphabetically. Third party imports were separated.